requirements.txt
networkx==2.5.1
r2pipe==1.5.3
rzpipe==0.1.0
psutil==5.7.3
pyaml==21.8.3